requirements.txt
black==23.7.0
bottle==0.12.25
click==8.1.6
mypy-extensions==1.0.0
packaging==23.1
Paste==3.5.3
pathspec==0.11.2
platformdirs==3.10.0
PyJWT==2.8.0
six==1.16.0
